Split trip controller
