+++ src/main/java/pt/ua/deti/tqs/backend/controllers/TripController.java
@@ -7,10 +7,8 @@ import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pt.ua.deti.tqs.backend.entities.Reservation;
 import pt.ua.deti.tqs.backend.entities.Trip;
 import pt.ua.deti.tqs.backend.helpers.Currency;
-import pt.ua.deti.tqs.backend.services.ReservationService;
 import pt.ua.deti.tqs.backend.services.TripService;
 import pt.ua.deti.tqs.backend.specifications.trip.TripSearchParameters;
 
@@ -18,18 +16,11 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController
-@RequestMapping("api/trip")
+@RequestMapping("public/trip")
 @Tag(name = "Trip")
 @AllArgsConstructor
 public class TripController {
     private final TripService tripService;
-    private final ReservationService reservationService;
-
-    @PostMapping
-    @Operation(summary = "Create a new trip")
-    public ResponseEntity<Trip> createTrip(@RequestBody Trip trip) {
-        return new ResponseEntity<>(tripService.createTrip(trip), HttpStatus.CREATED);
-    }
 
     @GetMapping
     @Operation(summary = "Get trips")
@@ -53,34 +44,5 @@ public class TripController {
         Trip trip = tripService.getTrip(id, currency);
         HttpStatus status = trip != null ? HttpStatus.OK : HttpStatus.NOT_FOUND;
         return new ResponseEntity<>(trip, status);
-    }
-
-    @GetMapping("{id}/reservations")
-    @Operation(summary = "Get all reservations of a trip")
-    public ResponseEntity<List<Reservation>> getReservationsByTripId(
-            @PathVariable @Parameter(name = "Trip ID", example = "1") Long id,
-            @RequestParam(required = false) @Parameter(name = "Currency", example = "EUR") Currency currency) {
-        List<Reservation> reservations = reservationService.getReservationsByTripId(id, currency);
-        HttpStatus status = !reservations.isEmpty() ? HttpStatus.OK : HttpStatus.NOT_FOUND;
-        return new ResponseEntity<>(reservations, status);
-    }
-
-    @PutMapping("{id}")
-    @Operation(summary = "Update a trip")
-    public ResponseEntity<Trip> updateTrip(
-            @PathVariable @Parameter(name = "Trip ID", example = "1") Long id,
-            @RequestBody Trip trip) {
-        trip.setId(id);
-        Trip updated = tripService.updateTrip(trip);
-        HttpStatus status = updated != null ? HttpStatus.OK : HttpStatus.NOT_FOUND;
-        return new ResponseEntity<>(updated, status);
-    }
-
-    @DeleteMapping("{id}")
-    @Operation(summary = "Delete a trip")
-    public ResponseEntity<Void> deleteTrip(
-            @PathVariable @Parameter(name = "Trip ID", example = "1") Long id) {
-        tripService.deleteTrip(id);
-        return new ResponseEntity<>(HttpStatus.OK);
     }
 }
